Add puppet command to failure details
Now the Reporter needs to know about the Puppeteer as well.

I'm not sure how I feel about a method called `command` returning an
array and consumers of that method needing to know to join it. On the
other hand, I don't really want to have a method called
`command_string`. It's kind of a dumb name.

It also makes the mock objects used in the tests more hairy, but those
should be ripped out and replaced with actual objects soon enough.