Actually set ENV vars in tests
I'll allow mock/stubbing on a lot of things, but I draw the line with
`ENV.expects(:[]).with('THIS').returns('that')`. Let's not be crazy.

Not that this pattern is so very much better. It's just actually
setting the values and using them, like the code is supposed to do.